+++ package.json
@@ -12,7 +12,7 @@
     "chart.js": "^2.9.3",
     "csvtojson": "^2.0.10",
     "data-pivot": "^1.0.2",
-    "easy-enums": "^1.1.7",
+    "easy-enums": "^1.1.8",
     "fast-json-patch": "^3.0.0-1",
     "hashmap": "^2.4.0",
     "intl": "^1.2.5",
@@ -20,7 +20,7 @@
     "js-cookie": "^2.2.1",
     "lodash": "^4.17.20",
     "md5": "^2.2.1",
-    "promise-value": "^1.2.9",
+    "promise-value": "^1.2.10",
     "react": "^16.13.1",
     "react-autocomplete": "^1.8.1",
     "react-chartjs2": "^1.2.1",
