+++ src/js/components/NewTabOnboarding.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Login from '../base/you-again';
+import Login from '../base/youagain';
 import NavBar, { AccountMenu } from './MyLoopNavBar';
 
 const NewTabOnboardingPage = () => {
