+++ src/js/components/NewtabLoginWidget.jsx
@@ -4,7 +4,7 @@
  */
 
 import React, { useState, useEffect } from 'react';
-import Login from '../base/you-again';
+import Login from '../base/youagain';
 import { assert, assMatch } from '../base/utils/assert';
 import C from '../C';
 import { emailLogin } from '../base/components/LoginWidget';
