+++ src/js/components/campaignpage/CampaignPage.jsx
@@ -2,7 +2,7 @@
  * 
  */
 import React, { Fragment, useState } from 'react';
-import Login from '../base/you-again';
+import Login from '../../base/youagain';
 import _ from 'lodash';
 import { Container, Alert } from 'reactstrap';
 import pivot from 'data-pivot';
