+++ src/js/components/cards/DonationCard.jsx
@@ -5,7 +5,7 @@ import ServerIO from '../../plumbing/ServerIO';
 import DataStore from '../../base/plumbing/DataStore';
 import Misc from '../../base/components/Misc';
 import ActionMan from '../../plumbing/ActionMan';
-import Login from '../base/you-again';
+import Login from '../base/youagain';
 
 /**
  * @returns {?Boolean} true if do-not-track is on
