+++ src/js/components/cards/RecentCampaignsCard.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import Login from '../base/you-again';
 import { isPortraitMobile, space } from '../../base/utils/miscutils';
 import { Col, Row, Alert } from 'reactstrap';
 import GoodLoopUnit from '../../base/components/GoodLoopUnit';
